requirements.txt
opencv_python==4.10.0.84
pymongo==4.8.0
python-dotenv==1.0.1
